+++ product-service/src/test/java/com/microservices/estore/product/ProductServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.microservices.estore.product_service;
+package com.microservices.estore.product;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
